src/requirements-root.txt
Keras
keras_metrics
jupyter
matplotlib
scipy
Pillow
Tensorflow
scikit-learn
pandas
datapunt-objectstore
